backend_todo_app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
openai==1.3.7
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
psycopg2-binary
python-dotenv==1.0.0